Separate planetary system components
