+++ src/routes/Home.tsx
@@ -2,7 +2,7 @@ export function Home({ }) {
     return (
         <>
             <div className='flex flex-col gap-4 text-gray-300 w-full'>
-                <h1 className='text-center text-7xl max-w-lg mx-auto font-bold text-white'>
+                <h1 className='text-center text-7xl max-w-xl mx-auto font-bold text-white'>
                     Powerfull <span className='text-transparent bg-clip-text bg-gradient-to-r from-orange-400 to-red-600'>tools</span> for the modern <i className='text-transparent bg-clip-text bg-gradient-to-r from-sky-400 to-green-400'>world builder</i>!</h1>
                 <p className='max-w-prose mx-auto text-center'>
                     Lorem, ipsum dolor sit amet consectetur adipisicing elit. Adipisci itaque distinctio a atque voluptates iste inventore vitae, numquam error exercitationem accusamus aliquam veritatis molestias repellendus? Sunt saepe inventore voluptatibus provident.
